feat: use ENABLE_POD_DESTROY environment variable
* Use ENABLE_POD_DESTROY environment variable (false by default)
* Update it using `kubectl patch` as part of the tutorial
