+++ README.md
@@ -123,7 +123,20 @@ Visit: **[http://localhost:3000](http://localhost:3000)**
 
 > **Podman users:** Replace `docker` with `podman` in the above commands.
 
-## 💡 Optional - Running poddestroy.js by setting docker-env --build-arg
+## 💡 Optional - Enable pod destruction!
 
-eval $(minikube -p minikube docker-env)
-docker build -t local/kubechaos:v5  --build-arg APP=poddestroy.js image
+You can change the ENABLE_POD_DESTROY environment variable to "true" to enable the pod destruction surprise!
+
+```bash
+kubectl patch deployment kubechaos -p '{"spec":{"template":{"spec":{"containers":[{"name":"app","env":[{"name":"ENABLE_POD_DESTROY","value":"true"}]}]}}}}'
+```
+
+The pods should restart automatically.
+
+```bash
+# Tip: you can watch the pods get replaced in real-time
+kubectl get pods -w
+```
+
+Refresh the page until you see a red button with "DESTROY POD NOW". The pod is killed (which you will be
+able to see with the above command) and replaced automatically.
