+++ image/app.js
@@ -1,4 +1,3 @@
-import { createServer } from 'http';
 import os from 'os';
 import express from 'express';
 
@@ -6,7 +5,7 @@ const port = process.env.PORT || 3000;
 
 
 // Array of HTML fragments (no DOM manipulation needed)
-const surprises = [
+let surprises = [
   `<h2>🎯 Click the target!</h2>
    <div style="font-size:100px;cursor:pointer;" onclick="alert('You hit it! 🎉')">🎯</div>`,
 
@@ -76,6 +75,51 @@ const server = express();
 
 server.use(express.static('public'));
 
+if (process.env.ENABLE_POD_DESTROY === "true") {
+  // Array of HTML fragments (no DOM manipulation needed)
+ surprises.push(
+    `<h2>💣 Click to destroy!</h2>
+  <button onclick="destroyPod()" style="font-size:20px;padding:10px;background:red;color:white;border:none;cursor:pointer;">💀 DESTROY POD NOW 💀</button>
+  <script>
+    function destroyPod() {
+      document.body.innerHTML = '<div style="background:black;color:red;font-size:50px;text-align:center;padding-top:200px;">💀 DESTROYING POD... 💀</div>';
+      console.log('Sending destroy request...');
+      fetch('/destroy', {method: 'POST'})
+        .then(response => {
+          console.log('Destroy response:', response.status);
+          document.body.innerHTML = '<div style="background:black;color:red;font-size:50px;text-align:center;padding-top:200px;">💀 POD DESTROYED 💀</div>';
+        })
+        .catch(error => {
+          console.error('Destroy failed:', error);
+          // Fallback - still try to show destruction
+          document.body.innerHTML = '<div style="background:black;color:red;font-size:50px;text-align:center;padding-top:200px;">💀 POD DESTROYED 💀</div>';
+        });
+    }
+  </script>`);
+
+  server.post("/destroy", (req, res) => {
+    console.log("💀💀💀 DESTRUCTION ENDPOINT HIT! 💀💀💀");
+    res.status(200).json({ message: "Pod is being destroyed!" });
+    console.log("🔥 KILLING PROCESS NOW...");
+
+    // Multiple ways to ensure the process dies
+    setTimeout(() => {
+        console.log("💥 PROCESS.EXIT(1)");
+        process.exit(1);
+    }, 100);
+
+    setTimeout(() => {
+        console.log("💥 PROCESS.EXIT(143) - SIGTERM");
+        process.exit(143);
+    }, 200);
+
+    setTimeout(() => {
+        console.log("💥 THROWING UNCAUGHT EXCEPTION");
+        throw new Error("INTENTIONAL CRASH FOR KUBERNETES EXPERIMENT");
+    }, 300);
+});
+}
+
 server.get('/', (req, res) => {
   const randomSurprise = surprises[Math.floor(Math.random() * surprises.length)];
   res.set("Cache-Control", "no-cache, no-store");
@@ -83,28 +127,6 @@ server.get('/', (req, res) => {
   res.send(renderPage(randomSurprise));
 });
 
-server.post("/destroy", (req, res) => {
-  console.log("💀💀💀 DESTRUCTION ENDPOINT HIT! 💀💀💀");
-  res.status(200).json({ message: "Pod is being destroyed!" });
-  console.log("🔥 KILLING PROCESS NOW...");
-
-  // Multiple ways to ensure the process dies
-  setTimeout(() => {
-    console.log("💥 PROCESS.EXIT(1)");
-    process.exit(1);
-  }, 100);
-
-  setTimeout(() => {
-    console.log("💥 PROCESS.EXIT(143) - SIGTERM");
-    process.exit(143);
-  }, 200);
-
-  setTimeout(() => {
-    console.log("💥 THROWING UNCAUGHT EXCEPTION");
-    throw new Error("INTENTIONAL CRASH FOR KUBERNETES EXPERIMENT");
-  }, 300);
-});
-
-server.listen(3000, () => {
+server.listen(port, () => {
   console.log('Server is running on http://localhost:3000');
 });
